feat: triton integration
