+++ pyproject.toml
@@ -133,7 +133,7 @@ io-file = ["filetype"] # Currently use for from_sample
 io-json = ["pydantic<2"] # currently we don't have support for pydantic 2.0
 io-image = ["bentoml[io-file]", "Pillow"]
 io-pandas = ["pandas", "pyarrow"]
-triton = ["bentoml[grpc]", "tritonclient[grpc]"]
+triton = ["tritonclient[grpc]>=2.29.0"]
 grpc = [
   "protobuf",
   "grpcio",
@@ -229,8 +229,6 @@ exclude = '''
     | typings
     | src/bentoml/grpc/v1alpha1
     | src/bentoml/grpc/v1
-    | src/bentoml/_internal/integrations/triton/_generated_pb3
-    | src/bentoml/_internal/integrations/triton/_generated_pb4
     | tests/proto
     | grpc-client/thirdparty
     | grpc-client/bentoml
@@ -259,8 +257,6 @@ ignore-paths = [
   "src/bentoml/_internal/external_typing",
   "src/bentoml/grpc/v1alpha1",
   "src/bentoml/grpc/v1",
-  "src/bentoml/_internal/integrations/triton/_generated_pb3",
-  "src/bentoml/_internal/integrations/triton/_generated_pb4",
 ]
 disable = ["coerce-builtin", "no-absolute-import", "C", "R"]
 enable = ["c-extension-no-member"]
