+++ src/bentoml/_internal/runner/container.py
@@ -9,23 +9,13 @@ from simple_di import inject
 from simple_di import Provide
 
 from ..types import LazyType
-from ..utils import LazyLoader
 from ..configuration.containers import BentoMLContainer
 
 SingleType = t.TypeVar("SingleType")
 BatchType = t.TypeVar("BatchType")
 
 if t.TYPE_CHECKING:
-    import tritonclient.grpc as tritongrpcclient
-
     from .. import external_typing as ext
-else:
-    tritongrpcclient = LazyLoader(
-        "tritongrpcclient",
-        globals(),
-        "tritonclient.grpc",
-        exc_msg="tritonclient is required to use triton with BentoML. Install with 'pip install bentoml[triton]'.",
-    )
 
 
 class Payload(t.NamedTuple):
