+++ src/bentoml/_internal/runner/runner_handle/__init__.py
@@ -15,7 +15,6 @@ if TYPE_CHECKING:
     R = t.TypeVar("R")
     P = t.ParamSpec("P")
 
-
 logger = logging.getLogger(__name__)
